feat: Refactor Python client to introduce index configurability  (#90)
