+++ api/routers/index.py
@@ -4,6 +4,7 @@ from pydantic import BaseModel
 from starlette.responses import JSONResponse
 from typing import List, Dict, Any
 
+from core.extract.postgres import PostgresExtractor
 from core.search.client import Client
 from core.search.index import default_model_name
 
@@ -17,7 +18,7 @@ client = Client(
 
 
 class IndexCreatePayload(BaseModel):
-    name: str
+    index_name: str
 
 
 class RegisterNeuralSearchFieldsPayload(BaseModel):
@@ -25,28 +26,92 @@ class RegisterNeuralSearchFieldsPayload(BaseModel):
     fields: List[str]
 
 
-class InsertDocumentsPayload(BaseModel):
-    index_name: str
-    documents: List[Dict[str, Any]]
-
-
 class SearchPayload(BaseModel):
     index_name: str
     dsl: Dict[str, Any]
 
 
-@router.post(f"/{tag}/create", tags=[tag])
-async def create_index(payload: IndexCreatePayload) -> JSONResponse:
+class AddSourcePayload(BaseModel):
+    index_name: str
+    source_host: str
+    source_port: int
+    source_user: str
+    source_password: str
+    source_relation: str
+    source_primary_key: str
+    source_columns: List[str]
+    source_neural_columns: List[str] = []
+
+
+@router.get("/index/{index_name}", tags=[tag])
+async def get_index(index_name: str) -> JSONResponse:
     try:
-        client.create_index(payload.name)
+        client.get_index(index_name)
         return JSONResponse(
-            status_code=status.HTTP_200_OK,
-            content=f"Index {payload.name} created successfully",
+            status_code=status.HTTP_200_OK, content=f"Index {index_name} found"
         )
     except Exception as e:
         return JSONResponse(
             status_code=status.HTTP_400_BAD_REQUEST,
-            content=f"Failed to create index {payload.name}: {e}",
+            content=str(e),
+        )
+
+
+@router.post(f"/{tag}/create", tags=[tag])
+async def create_index(payload: IndexCreatePayload) -> JSONResponse:
+    try:
+        print("creating")
+        client.create_index(payload.index_name)
+        return JSONResponse(
+            status_code=status.HTTP_200_OK,
+            content=f"Index {payload.index_name} created successfully",
+        )
+    except Exception as e:
+        return JSONResponse(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            content=f"Failed to create index {payload.index_name}: {e}",
+        )
+
+
+@router.post(f"/{tag}/add_source", tags=[tag])
+async def add_source(payload: AddSourcePayload) -> JSONResponse:
+    # Number of rows to extract at once
+    BATCH_SIZE = 500
+
+    try:
+        index = client.get_index(payload.index_name)
+
+        extractor = PostgresExtractor(
+            host=payload.source_host,
+            port=payload.source_port,
+            user=payload.source_user,
+            password=payload.source_password,
+        )
+
+        if payload.source_neural_columns:
+            index.register_neural_search_fields(payload.source_neural_columns)
+
+        for chunk in extractor.extract_all(
+            relation=payload.source_relation,
+            columns=payload.source_columns,
+            primary_key=payload.source_primary_key,
+            chunk_size=BATCH_SIZE,
+        ):
+            rows = chunk.get("rows")
+            primary_keys = chunk.get("primary_keys")
+
+            if rows and primary_keys:
+                print("upserting", rows)
+                index.upsert(documents=rows, ids=primary_keys)
+
+        return JSONResponse(
+            status_code=status.HTTP_200_OK,
+            content=f"Source {payload.source_relation} linked to index {payload.index_name} successfully",
+        )
+    except Exception as e:
+        return JSONResponse(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            content=str(e),
         )
 
 
@@ -68,58 +133,13 @@ async def register_neural_search_fields(
         )
 
 
-@router.post(f"/{tag}/insert", tags=[tag])
-async def insert_documents(payload: InsertDocumentsPayload) -> JSONResponse:
-    try:
-        index = client.get_index(payload.index_name)
-        index.insert(payload.documents)
-        return JSONResponse(
-            status_code=status.HTTP_200_OK,
-            content=f"Documents inserted successfully into {payload.index_name}",
-        )
-    except Exception as e:
-        return JSONResponse(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            content=f"Failed to insert documents: {e}",
-        )
-
-
 @router.post(f"/{tag}/search", tags=[tag])
 async def search_documents(payload: SearchPayload) -> JSONResponse:
     try:
-
-        def add_model_id(nested_dict, model_id):
-            for key, value in nested_dict.items():
-                if isinstance(value, dict):
-                    if "source" not in value.keys():
-                        add_model_id(value, model_id)
-                    if key == "neural":
-                        for inner_key, inner_value in value.items():
-                            if (
-                                isinstance(inner_value, dict)
-                                and "source" not in inner_value.keys()
-                            ):
-                                inner_value["model_id"] = model_id
-                elif isinstance(value, list):
-                    for item in value:
-                        if isinstance(item, dict):
-                            add_model_id(item, model_id)
-
         index = client.get_index(payload.index_name)
-        dsl = payload.dsl
-        model = index.model.get(default_model_name)
-
-        if not model:
-            return JSONResponse(
-                status_code=status.HTTP_400_BAD_REQUEST,
-                content=f"Table {payload.index_name} is not properly registered. Did you forget to index() the table?",
-            )
-
-        model_id = model["model_id"]
-        add_model_id(dsl, model_id)
-        print(dsl)
-
-        return JSONResponse(status_code=status.HTTP_200_OK, content=index.search(dsl))
+        return JSONResponse(
+            status_code=status.HTTP_200_OK, content=index.search(payload.dsl)
+        )
     except Exception as e:
         return JSONResponse(
             status_code=status.HTTP_400_BAD_REQUEST,
