+++ clients/python/retakesearch/__init__.py
@@ -1,4 +1,5 @@
-from .client import Client, Database, Table
+from .client import Client
+from .index import Index, Database, Table
 from .search import Search, Q, A
 
-__all__ = ["Client", "Search", "Database", "Table", "Q", "A"]
+__all__ = ["Client", "Search", "Database", "Table", "Index", "Q", "A"]
