+++ docs/company/roadmap.mdx
@@ -6,7 +6,7 @@ description: Last Updated July 26, 2023
 The goal of this page is to provide transparency into what we're working on, what we've completed, and to align
 the efforts of the core Retake team with community efforts.
 
-### Upcoming Features
+## Upcoming Features
 
 Below is a list of high-priority features that the Retake team is working on right now.
 
@@ -22,9 +22,9 @@ Below is a list of high-priority features that the Retake team is working on rig
 The best way to request a feature is to [create a Github issue](https://github.com/getretake/retake/issues/new/choose)
 for it. We read these issues every day and are very responsive.
 
-### Completed Features
+## Completed Features
 
-A list of features that have been completed
+A list of features that have been completed.
 
 **July 25, 2023: First MVP**: We're excited to release our first MVP into the wild! It's very simple: just a Python SDK
 with two functions: one to index a Postgres table, and another to execute search queries with the full expressiveness
